test(cli): export Prisma 테스트를 tests 모듈로 분리
